seq: event_data_queue: fix type of argument for getter of byte param
Although the getter of byte param has argument of guint32 type, the
argument should be guint8 type.

Fixes: a272b5d0e727: ("seq: event_data_queue: add accessor methods")

+++ src/seq/event-data-queue.h
@@ -43,7 +43,7 @@ void alsaseq_event_data_queue_set_quadlet_param(ALSASeqEventDataQueue *self,
 
 const guint8 *alsaseq_event_data_queue_get_byte_param(ALSASeqEventDataQueue *self);
 void alsaseq_event_data_queue_set_byte_param(ALSASeqEventDataQueue *self,
-                                             const guint32 bytes[8]);
+                                             const guint8 bytes[8]);
 
 G_END_DECLS
